move to separate file and clean up
